## Summary
- Finish the BE/64-bit sweep: all remaining writers now use explicit big-endian helpers (db header fields, menu/lang packers, wpengine, pict length, ODB/Cancoon, shell sysverbs, memory serialization). Added dated Codex notes per file.
- Add procedural byte-level goldens in tests/db_format_tests.c (v7 header/avail encoding + PICT length BE32) and fix signed/unsigned warnings in oplist.c and db_format.c.
- Update planning/_CURRENT_STATUS.md and planning/big_endian_portability_audit.md to mark the sweep complete and note the procedural cross-arch sentinel.

## So what
- V7 output is now fully 64-bit and endian-agnostic, with byte-level tests guarding drift; runtime-only fields stay zeroed on disk; PICT payloads remain opaque.
- Procedural goldens provide cross-arch enforcement until x86 CI runs.

## Tests
- make -C tests db_format_tests
- make -C tests runtime_tests

+++ Common/source/db_format.c
@@ -1,3 +1,5 @@
+/* 2025-11-24 Codex: Clamp header comparisons with uint64 for BE safety. */
+
 #include "frontier.h"
 #include "standard.h"
 #include "shell_api.h"
@@ -1020,14 +1022,14 @@ boolean migrate_32bit_to_64bit(const char *db_path) {
     if (!tablesavesystemtable(hrootvariable, &new_root_address))
         goto cleanup;
 
-    if (new_root_address > 0xFFFFFFFFULL)
+    if ((uint64_t) new_root_address > 0xFFFFFFFFULL)
         goto cleanup;
 
     if (hscript != nil) {
         new_script_address = script_address;
         if (!dbassignhandle(hscript, &new_script_address))
             goto cleanup;
-        if (new_script_address > 0xFFFFFFFFULL)
+        if ((uint64_t) new_script_address > 0xFFFFFFFFULL)
             goto cleanup;
     } else {
         new_script_address = 0;
